Refactor SessionsController#destroy_others
