--- app/listeners/casino/other_sessions_destroyer_listener.rb
@@ -1,7 +1,0 @@
-require_relative 'listener'
-
-class CASino::OtherSessionsDestroyerListener < CASino::Listener
-  def other_sessions_destroyed(url)
-    @controller.redirect_to(url || sessions_path)
-  end
-end
